separated server into glamor-server (mhh naming?)
