+++ package.json
@@ -50,7 +50,6 @@
     "pegjs": "^0.10.0",
     "preact": "^7.1.0",
     "preact-compat": "^3.9.4",
-    "pug": "^2.0.0-beta6",
     "react": "^15.4.1",
     "react-addons-css-transition-group": "^15.4.1",
     "react-dom": "^15.4.1",
